# handler/api.gen.go
// Package handler provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package handler

import (
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"

	"github.com/gin-gonic/gin"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

const (
	CookieAuthScopes = "cookieAuth.Scopes"
)

// AcknowledgeRequestPayload defines model for AcknowledgeRequestPayload.
type AcknowledgeRequestPayload struct {
	// Message Message for the acknowledgment request
	Message string `json:"message"`

	// ToUsernames List of usernames to send acknowledgment request to
	ToUsernames []string `json:"to_usernames"`
}

// AcknowledgeRequestResponse defines model for AcknowledgeRequestResponse.
type AcknowledgeRequestResponse struct {
	RequestId *string `json:"request_id,omitempty"`
	Success   *bool   `json:"success,omitempty"`
}

// AcknowledgeResponsePayload defines model for AcknowledgeResponsePayload.
type AcknowledgeResponsePayload struct {
	// RequestId ID of the acknowledgment request being responded to
	RequestId string `json:"request_id"`
}

// AcknowledgeResponseResponse defines model for AcknowledgeResponseResponse.
type AcknowledgeResponseResponse struct {
	Success *bool `json:"success,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Username string `json:"username"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	Success  *bool   `json:"success,omitempty"`
	Username *string `json:"username,omitempty"`
}

// LogoutResponse defines model for LogoutResponse.
type LogoutResponse struct {
	Success *bool `json:"success,omitempty"`
}

// NotifyRequest defines model for NotifyRequest.
type NotifyRequest struct {
	FromUsername string `json:"from_username"`
	Message      string `json:"message"`

	// TargetUsername A specific username or 'all'
	TargetUsername string `json:"target_username"`
}

// NotifyResponse defines model for NotifyResponse.
type NotifyResponse struct {
	Success *bool `json:"success,omitempty"`
}

// UsersResponse defines model for UsersResponse.
type UsersResponse struct {
	// Users List of connected usernames
	Users *[]string `json:"users,omitempty"`
}

// PostAcknowledgeRequestJSONRequestBody defines body for PostAcknowledgeRequest for application/json ContentType.
type PostAcknowledgeRequestJSONRequestBody = AcknowledgeRequestPayload

// PostAcknowledgeResponseJSONRequestBody defines body for PostAcknowledgeResponse for application/json ContentType.
type PostAcknowledgeResponseJSONRequestBody = AcknowledgeResponsePayload

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = LoginRequest

// PostNotifyJSONRequestBody defines body for PostNotify for application/json ContentType.
type PostNotifyJSONRequestBody = NotifyRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Sends an acknowledgment request to user(s) (requires authentication)
	// (POST /acknowledge/request)
	PostAcknowledgeRequest(c *gin.Context)
	// Responds to an acknowledgment request (requires authentication)
	// (POST /acknowledge/response)
	PostAcknowledgeResponse(c *gin.Context)
	// Subscribes to the SSE notification stream (requires authentication)
	// (GET /events)
	GetEvents(c *gin.Context)
	// Logs a user in and creates a session
	// (POST /login)
	PostLogin(c *gin.Context)
	// Logs a user out and destroys the session
	// (POST /logout)
	PostLogout(c *gin.Context)
	// Broadcasts a notification (requires authentication)
	// (POST /notify)
	PostNotify(c *gin.Context)
	// Gets list of currently connected users (requires authentication)
	// (GET /users)
	GetUsers(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostAcknowledgeRequest operation middleware
func (siw *ServerInterfaceWrapper) PostAcknowledgeRequest(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAcknowledgeRequest(c)
}

// PostAcknowledgeResponse operation middleware
func (siw *ServerInterfaceWrapper) PostAcknowledgeResponse(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAcknowledgeResponse(c)
}

// GetEvents operation middleware
func (siw *ServerInterfaceWrapper) GetEvents(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetEvents(c)
}

// PostLogin operation middleware
func (siw *ServerInterfaceWrapper) PostLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostLogin(c)
}

// PostLogout operation middleware
func (siw *ServerInterfaceWrapper) PostLogout(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostLogout(c)
}

// PostNotify operation middleware
func (siw *ServerInterfaceWrapper) PostNotify(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostNotify(c)
}

// GetUsers operation middleware
func (siw *ServerInterfaceWrapper) GetUsers(c *gin.Context) {

	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetUsers(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/acknowledge/request", wrapper.PostAcknowledgeRequest)
	router.POST(options.BaseURL+"/acknowledge/response", wrapper.PostAcknowledgeResponse)
	router.GET(options.BaseURL+"/events", wrapper.GetEvents)
	router.POST(options.BaseURL+"/login", wrapper.PostLogin)
	router.POST(options.BaseURL+"/logout", wrapper.PostLogout)
	router.POST(options.BaseURL+"/notify", wrapper.PostNotify)
	router.GET(options.BaseURL+"/users", wrapper.GetUsers)
}

type PostAcknowledgeRequestRequestObject struct {
	Body *PostAcknowledgeRequestJSONRequestBody
}

type PostAcknowledgeRequestResponseObject interface {
	VisitPostAcknowledgeRequestResponse(w http.ResponseWriter) error
}

type PostAcknowledgeRequest200JSONResponse AcknowledgeRequestResponse

func (response PostAcknowledgeRequest200JSONResponse) VisitPostAcknowledgeRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostAcknowledgeRequest400Response struct {
}

func (response PostAcknowledgeRequest400Response) VisitPostAcknowledgeRequestResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostAcknowledgeRequest401Response struct {
}

func (response PostAcknowledgeRequest401Response) VisitPostAcknowledgeRequestResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostAcknowledgeResponseRequestObject struct {
	Body *PostAcknowledgeResponseJSONRequestBody
}

type PostAcknowledgeResponseResponseObject interface {
	VisitPostAcknowledgeResponseResponse(w http.ResponseWriter) error
}

type PostAcknowledgeResponse200JSONResponse AcknowledgeResponseResponse

func (response PostAcknowledgeResponse200JSONResponse) VisitPostAcknowledgeResponseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostAcknowledgeResponse400Response struct {
}

func (response PostAcknowledgeResponse400Response) VisitPostAcknowledgeResponseResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostAcknowledgeResponse401Response struct {
}

func (response PostAcknowledgeResponse401Response) VisitPostAcknowledgeResponseResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type GetEventsRequestObject struct {
}

type GetEventsResponseObject interface {
	VisitGetEventsResponse(w http.ResponseWriter) error
}

type GetEvents200TexteventStreamResponse struct {
	Body          io.Reader
	ContentLength int64
}

func (response GetEvents200TexteventStreamResponse) VisitGetEventsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "text/event-stream")
	if response.ContentLength != 0 {
		w.Header().Set("Content-Length", fmt.Sprint(response.ContentLength))
	}
	w.WriteHeader(200)

	if closer, ok := response.Body.(io.ReadCloser); ok {
		defer closer.Close()
	}
	_, err := io.Copy(w, response.Body)
	return err
}

type GetEvents401Response struct {
}

func (response GetEvents401Response) VisitGetEventsResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostLoginRequestObject struct {
	Body *PostLoginJSONRequestBody
}

type PostLoginResponseObject interface {
	VisitPostLoginResponse(w http.ResponseWriter) error
}

type PostLogin200ResponseHeaders struct {
	SetCookie string
}

type PostLogin200JSONResponse struct {
	Body    LoginResponse
	Headers PostLogin200ResponseHeaders
}

func (response PostLogin200JSONResponse) VisitPostLoginResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Set-Cookie", fmt.Sprint(response.Headers.SetCookie))
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response.Body)
}

type PostLogin400Response struct {
}

func (response PostLogin400Response) VisitPostLoginResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostLogin401Response struct {
}

func (response PostLogin401Response) VisitPostLoginResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostLogoutRequestObject struct {
}

type PostLogoutResponseObject interface {
	VisitPostLogoutResponse(w http.ResponseWriter) error
}

type PostLogout200JSONResponse LogoutResponse

func (response PostLogout200JSONResponse) VisitPostLogoutResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostLogout401Response struct {
}

func (response PostLogout401Response) VisitPostLogoutResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostNotifyRequestObject struct {
	Body *PostNotifyJSONRequestBody
}

type PostNotifyResponseObject interface {
	VisitPostNotifyResponse(w http.ResponseWriter) error
}

type PostNotify200JSONResponse NotifyResponse

func (response PostNotify200JSONResponse) VisitPostNotifyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostNotify400Response struct {
}

func (response PostNotify400Response) VisitPostNotifyResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostNotify401Response struct {
}

func (response PostNotify401Response) VisitPostNotifyResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type GetUsersRequestObject struct {
}

type GetUsersResponseObject interface {
	VisitGetUsersResponse(w http.ResponseWriter) error
}

type GetUsers200JSONResponse UsersResponse

func (response GetUsers200JSONResponse) VisitGetUsersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetUsers401Response struct {
}

func (response GetUsers401Response) VisitGetUsersResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Sends an acknowledgment request to user(s) (requires authentication)
	// (POST /acknowledge/request)
	PostAcknowledgeRequest(ctx context.Context, request PostAcknowledgeRequestRequestObject) (PostAcknowledgeRequestResponseObject, error)
	// Responds to an acknowledgment request (requires authentication)
	// (POST /acknowledge/response)
	PostAcknowledgeResponse(ctx context.Context, request PostAcknowledgeResponseRequestObject) (PostAcknowledgeResponseResponseObject, error)
	// Subscribes to the SSE notification stream (requires authentication)
	// (GET /events)
	GetEvents(ctx context.Context, request GetEventsRequestObject) (GetEventsResponseObject, error)
	// Logs a user in and creates a session
	// (POST /login)
	PostLogin(ctx context.Context, request PostLoginRequestObject) (PostLoginResponseObject, error)
	// Logs a user out and destroys the session
	// (POST /logout)
	PostLogout(ctx context.Context, request PostLogoutRequestObject) (PostLogoutResponseObject, error)
	// Broadcasts a notification (requires authentication)
	// (POST /notify)
	PostNotify(ctx context.Context, request PostNotifyRequestObject) (PostNotifyResponseObject, error)
	// Gets list of currently connected users (requires authentication)
	// (GET /users)
	GetUsers(ctx context.Context, request GetUsersRequestObject) (GetUsersResponseObject, error)
}

type StrictHandlerFunc = strictgin.StrictGinHandlerFunc
type StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PostAcknowledgeRequest operation middleware
func (sh *strictHandler) PostAcknowledgeRequest(ctx *gin.Context) {
	var request PostAcknowledgeRequestRequestObject

	var body PostAcknowledgeRequestJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostAcknowledgeRequest(ctx, request.(PostAcknowledgeRequestRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostAcknowledgeRequest")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostAcknowledgeRequestResponseObject); ok {
		if err := validResponse.VisitPostAcknowledgeRequestResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostAcknowledgeResponse operation middleware
func (sh *strictHandler) PostAcknowledgeResponse(ctx *gin.Context) {
	var request PostAcknowledgeResponseRequestObject

	var body PostAcknowledgeResponseJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostAcknowledgeResponse(ctx, request.(PostAcknowledgeResponseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostAcknowledgeResponse")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostAcknowledgeResponseResponseObject); ok {
		if err := validResponse.VisitPostAcknowledgeResponseResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetEvents operation middleware
func (sh *strictHandler) GetEvents(ctx *gin.Context) {
	var request GetEventsRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetEvents(ctx, request.(GetEventsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetEvents")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetEventsResponseObject); ok {
		if err := validResponse.VisitGetEventsResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostLogin operation middleware
func (sh *strictHandler) PostLogin(ctx *gin.Context) {
	var request PostLoginRequestObject

	var body PostLoginJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostLogin(ctx, request.(PostLoginRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostLogin")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostLoginResponseObject); ok {
		if err := validResponse.VisitPostLoginResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostLogout operation middleware
func (sh *strictHandler) PostLogout(ctx *gin.Context) {
	var request PostLogoutRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostLogout(ctx, request.(PostLogoutRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostLogout")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostLogoutResponseObject); ok {
		if err := validResponse.VisitPostLogoutResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostNotify operation middleware
func (sh *strictHandler) PostNotify(ctx *gin.Context) {
	var request PostNotifyRequestObject

	var body PostNotifyJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostNotify(ctx, request.(PostNotifyRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostNotify")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostNotifyResponseObject); ok {
		if err := validResponse.VisitPostNotifyResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetUsers operation middleware
func (sh *strictHandler) GetUsers(ctx *gin.Context) {
	var request GetUsersRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetUsers(ctx, request.(GetUsersRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetUsers")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetUsersResponseObject); ok {
		if err := validResponse.VisitGetUsersResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}